pyproject.toml
[project]
name = "ai-blog-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.129.0",
    "openai>=2.21.0",
    "openai-agents>=0.9.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "rich>=14.3.3",
    "uvicorn>=0.41.0",
]
